parsing output in views with .html_safe
