+++ app/views/tweets/index.html.slim
@@ -20,6 +20,6 @@
     tbody
       - @tweets.each do |tweet|
         tr
-          td = tweet.message
+          td = tweet.message.html_safe
           td = link_to tweet.user.username, user_path(tweet.user)
           td = tweet.created_at
