+++ config/routes.rb
@@ -7,4 +7,5 @@ Rails.application.routes.draw do
 
   resources :users, only: [:new, :create, :show]
   resources :tweets, only: [:index, :create]
+  resources :hashtags, only: :show
 end
